+++ Package.swift
@@ -20,12 +20,13 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/cezheng/Fuzi.git", from: "3.1.3"),
-        .package(url: "https://github.com/krzyzanowskim/CryptoSwift.git", from: "1.3.8"),
-        .package(url: "https://github.com/marmelroy/Zip.git", from: "2.1.1"),
-        .package(url: "https://github.com/ra1028/DifferenceKit.git", from: "1.2.0"),
-        .package(url: "https://github.com/readium/GCDWebServer.git", from: "3.7.0"),
-        .package(url: "https://github.com/scinfu/SwiftSoup.git", from: "2.3.2"),
-        .package(url: "https://github.com/stephencelis/SQLite.swift.git", from: "0.12.2"),
+        // From 1.6.0, the build fails in GitHub actions with Carthage
+        .package(url: "https://github.com/krzyzanowskim/CryptoSwift.git", "1.5.1"..<"1.6.0"),
+        .package(url: "https://github.com/marmelroy/Zip.git", from: "2.1.2"),
+        .package(url: "https://github.com/ra1028/DifferenceKit.git", from: "1.3.0"),
+        .package(url: "https://github.com/readium/GCDWebServer.git", from: "3.7.3"),
+        .package(url: "https://github.com/scinfu/SwiftSoup.git", from: "2.4.3"),
+        .package(url: "https://github.com/stephencelis/SQLite.swift.git", from: "0.13.3"),
         // 0.9.12 requires iOS 12+
         .package(url: "https://github.com/weichsel/ZIPFoundation.git", "0.9.0"..<"0.9.12"),
     ],